% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_metrics.R
\name{calibrate_metrics}
\alias{calibrate_metrics}
\alias{calibrate_metrics.data.frame}
\alias{calibrate_metrics.numeric}
\title{Calculates a set of metrics}
\usage{
calibrate_metrics(object, ...)

\method{calibrate_metrics}{data.frame}(object, resid_col, metrics, ...)

\method{calibrate_metrics}{numeric}(object, metrics, ...)
}
\arguments{
\item{object}{a data.frame or numeric vector containing residuals}

\item{...}{unused.}

\item{resid_col}{column name corresponding to residuals.}

\item{metrics}{a list of \code{lanyard} metrics.}
}
\description{
Calculates a set of metrics
}
