% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhsic_sq_metric.R
\name{dhsic_sq_metric}
\alias{dhsic_sq_metric}
\alias{dhsic_sq_metric.data.frame}
\alias{dhsic_sq_metric_vec}
\alias{dhsic_sq_metric.numeric}
\title{Multiavariate Hilbert Schmidt Independence Criterion metric}
\usage{
dhsic_sq_metric(data, ...)

\method{dhsic_sq_metric}{data.frame}(
  data,
  truth,
  ...,
  lags = 2:vctrs::vec_size(data) - 1,
  na_rm = TRUE,
  case_weights = NULL
)

dhsic_sq_metric_vec(
  truth,
  estimate,
  lags = 0:1,
  na_rm = TRUE,
  case_weights = NULL,
  ...
)

\method{dhsic_sq_metric}{numeric}(data, lags = 1, case_weights = NULL, ...)
}
\arguments{
\item{data}{A \code{data.frame} containing the columns specified by \code{truth} and
\code{...}.}

\item{...}{A set of unquoted column names or one or more
\code{dplyr} selector functions to choose which variables contain the
class probabilities. If \code{truth} is binary, only 1 column should be selected,
and it should correspond to the value of \code{event_level}. Otherwise, there
should be as many columns as factor levels of \code{truth} and the ordering of
the columns should be the same as the factor levels of \code{truth}.}

\item{truth}{The column identifier for the true results (that is \code{numeric}).}

\item{lags}{Vector of positive integers. This corresponds to the lags at
which distance covariance should be evaluated.}

\item{na_rm}{A \code{logical} value indicating whether \code{NA}
values should be stripped before the computation proceeds.}

\item{case_weights}{The optional column identifier for case weights.
This should be an unquoted column name that evaluates to a numeric column
in \code{data}. For \verb{_vec()} functions, a numeric vector.}

\item{estimate}{The column identifier for the predicted results (that is also
\code{numeric})}
}
\value{
A \code{tibble} of class \code{dhsic_tbl} with columns \code{lag}, \code{dhsic} and
\code{dhsic_scaled}.
}
\description{
Multiavariate Hilbert Schmidt Independence Criterion metric
}
\examples{
# dhsic ---------------------------------------------------------------------
data <- data.frame(t = rnorm(100), e = rnorm(100))
dhsic_sq_metric(data, t, e)
}
