% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lanyard-package.R
\docType{package}
\name{lanyard-package}
\alias{lanyard}
\alias{lanyard-package}
\title{lanyard: Lag Dependence Yardstick Metrics}
\description{
This package makes available serial dependence metrics including the autocorrelation, partial autocorrelation and auto-distance correlation. These metrics are registered with the `yardstick` package.
}
\author{
\strong{Maintainer}: Leon Fernandes \email{leon.fernandes@columbia.edu} (\href{https://orcid.org/0000-0002-7192-9939}{ORCID})

}
\keyword{internal}
