% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-srl_dep_tbl.R
\name{tidy.srl_dep_tbl}
\alias{tidy.srl_dep_tbl}
\title{Tidy method for serial dependence}
\usage{
tidy.srl_dep_tbl(x, ...)
}
\arguments{
\item{x}{a \code{tibble} of class \code{srl_dep_tbl}.}

\item{...}{Additional arguments to tidying method.}
}
\value{
a \code{tibble} with columns \code{.metric}, \code{.estimator}, \code{lag} and
\code{.estimate}.
}
\description{
Tidy method for serial dependence
}
