% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pacf_metric.R
\name{pacf_metric}
\alias{pacf_metric}
\alias{pacf_metric.data.frame}
\alias{pacf_metric_vec}
\title{Partial autocorrelation metric}
\usage{
pacf_metric(data, ...)

\method{pacf_metric}{data.frame}(
  data,
  truth,
  ...,
  lag = 2:vctrs::vec_size(data) - 1,
  na_rm = TRUE,
  case_weights = NULL
)

pacf_metric_vec(
  truth,
  estimate,
  lag = 0:1,
  na_rm = TRUE,
  case_weights = NULL,
  ...
)
}
\arguments{
\item{data}{A \code{data.frame} containing the columns specified by \code{truth} and
\code{...}.}

\item{...}{A set of unquoted column names or one or more
\code{dplyr} selector functions to choose which variables contain the
class probabilities. If \code{truth} is binary, only 1 column should be selected,
and it should correspond to the value of \code{event_level}. Otherwise, there
should be as many columns as factor levels of \code{truth} and the ordering of
the columns should be the same as the factor levels of \code{truth}.}

\item{truth}{The column identifier for the true results (that is \code{numeric}).}

\item{lag}{Vector of positive integers. This corresponds to the lag at
which distance covariance should be evaluated.}

\item{na_rm}{A \code{logical} value indicating whether \code{NA}
values should be stripped before the computation proceeds.}

\item{case_weights}{The optional column identifier for case weights.
This should be an unquoted column name that evaluates to a numeric column
in \code{data}. For \verb{_vec()} functions, a numeric vector.}

\item{estimate}{The column identifier for the predicted results (that is also
\code{numeric})}
}
\value{
A \code{tibble} with columns \code{.metric}, \code{.estimator} and \code{.estimate} and
1 row of values. For \code{pacf_metric_vec}, a single \code{numeric} value (or \code{NA}).
}
\description{
Partial autocorrelation metric
}
